refactoring the code into a setuptools
