--- components/__init__.py
@@ -1,2 +1,0 @@
-from components.ip_lookup import ip_lookup
-from components.repositories import repo_check_incoming
