+++ src/pybliotecario/configurationDataTemplate.py
