+++ src/pybliotecario/on_call.py
@@ -1,5 +1,5 @@
-from configurationData import chatId
-from TelegramUtil import TelegramUtil
+from pybliotecario.configurationData import chatId
+from pybliotecario.TelegramUtil import TelegramUtil
 import pdb
 
 
@@ -9,20 +9,20 @@ def run_command(args):
 
     if args.pid:
         """ Wait until the given pid finish, then do whatever else you have been told to do """
-        from components.pid import wait_for_it_until_finished
+        from pybliotecario.components.pid import wait_for_it_until_finished
         print("Waiting for the given PIDs: {0}".format(args.pid))
         wait_for_it_until_finished(args.pid)
 
     if args.arxiv_new:
-        from components.arxiv_functions import arxiv_recent_filtered
-        from configurationData import arxiv_filter_dict, arxiv_categories
+        from pybliotecario.components.arxiv_functions import arxiv_recent_filtered
+        from pybliotecario.configurationData import arxiv_filter_dict, arxiv_categories
         msg = arxiv_recent_filtered(arxiv_categories, arxiv_filter_dict)
         teleAPI.send_message(msg, chatId)
         print("Arxiv information sent")
 
     if args.weather:
-        from components.weather import will_it_rain, check_current_weather
-        from configurationData import weather_api, weather_location, weather_times
+        from pybliotecario.components.weather import will_it_rain, check_current_weather
+        from pybliotecario.configurationData import weather_api, weather_location, weather_times
         msg_rain = will_it_rain(weather_location, weather_api, weather_times)
         msg_wthr = check_current_weather(weather_location, weather_api)
         weather_msg = "{1}\n{0}".format(msg_rain, msg_wthr)
@@ -30,7 +30,7 @@ def run_command(args):
         print("Weather information sent")
 
     if args.check_repository:
-        from components import repo_check_incoming
+        from pybliotecario.components import repo_check_incoming
         msg = repo_check_incoming(args.check_repository)
         teleAPI.send_message(msg, chatId)
 
@@ -44,7 +44,7 @@ def run_command(args):
 
 
     if args.my_ip:
-        from components.ip_lookup import ip_lookup
+        from pybliotecario.components.ip_lookup import ip_lookup
         my_ip = ip_lookup(None, force = True)
         message_text = " ".join(args.message)
         message = f'{message_text} {my_ip}'
